Conflicts:
	src/main/webapp/WEB-INF/views/users.html

+++ src/main/java/org/telosys/tools/helper/FileUnit.java
@@ -5,6 +5,6 @@ package org.telosys.tools.helper;
  */
 public class FileUnit {
 
-    public final static long MEGABYTE = 1024*1024;
+    public final static double MEGABYTE = 1024*1024;
 
 }
